preparation/requirements.txt
psd-tools>=1.9.0
Pillow>=10.0.0
numpy>=1.24.0
google-genai>=0.3.0
python-dotenv>=1.0.0
tqdm>=4.67.1